Move crash code to its own file.
