+++ src/main.cpp
@@ -9,6 +9,7 @@
 #include "google/protobuf/stubs/common.h"
 #include "version.hpp"
 #include "utility.hpp"
+#include "crash.hpp"
 
 // New subcommand system provides all the subcommands that used to live here
 #include "subcommand/subcommand.hpp"
